Fix for server held item slot

+++ ac.artemis.packet.spigot/src/main/java/cc/ghast/packet/wrapper/packet/play/server/GPacketPlayServerHeldItemSlot.java
@@ -2,9 +2,9 @@ package cc.ghast.packet.wrapper.packet.play.server;
 
 import ac.artemis.packet.protocol.ProtocolVersion;
 import ac.artemis.packet.spigot.protocol.PacketLink;
+import ac.artemis.packet.spigot.wrappers.GPacket;
 import ac.artemis.packet.wrapper.server.PacketPlayServerItemHeldSlot;
 import cc.ghast.packet.buffer.ProtocolByteBuf;
-import ac.artemis.packet.spigot.wrappers.GPacket;
 import cc.ghast.packet.wrapper.packet.ReadableBuffer;
 import cc.ghast.packet.wrapper.packet.WriteableBuffer;
 import lombok.Getter;
@@ -15,13 +15,13 @@ import java.util.UUID;
 @PacketLink(PacketPlayServerItemHeldSlot.class)
 public class GPacketPlayServerHeldItemSlot extends GPacket implements PacketPlayServerItemHeldSlot, ReadableBuffer, WriteableBuffer {
 
-    private short slot;
+    private byte slot;
 
     public GPacketPlayServerHeldItemSlot(UUID player, ProtocolVersion version) {
         super("PacketPlayOutHeldItemSlot", player, version);
     }
 
-    public GPacketPlayServerHeldItemSlot(short slot) {
+    public GPacketPlayServerHeldItemSlot(byte slot) {
         super("PacketPlayOutHeldItemSlot");
 
         this.slot = slot;
@@ -29,11 +29,11 @@ public class GPacketPlayServerHeldItemSlot extends GPacket implements PacketPlay
 
     @Override
     public void read(ProtocolByteBuf byteBuf) {
-        this.slot = byteBuf.readShort();
+        this.slot = byteBuf.readByte();
     }
 
     @Override
     public void write(ProtocolByteBuf byteBuf) {
-        byteBuf.writeShort(this.slot);
+        byteBuf.writeByte(this.slot);
     }
 }
